requirements.txt
decorator==4.0.2
Flask==0.10.1
Flask-PyMongo==0.3.1
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.8
libvirt-python==1.2.18
MarkupSafe==0.23
pbr==1.6.0
pymongo==2.8.1
sh==1.11
six==1.9.0
SQLAlchemy==1.0.8
sqlalchemy-migrate==0.9.7
sqlparse==0.1.16
Tempita==0.5.2
Werkzeug==0.10.4
wheel==0.24.0
